Moved RateLimiter.scala (without any CircuitBreaker mind you) back to Chapter12 as per book Fixed sbt.IO dependency in build.scala and chapter13 (must have been in .ivy cache for RKuhn), removed some unused imports directory change for MultiMasterCRDTSpec.scala (now chapter13)
