+++ project/Build.scala
@@ -17,7 +17,7 @@ object Build {
 
   val amazonAWS = "com.amazonaws" % "aws-java-sdk" % "1.11.109"
 
-  val sbtIO = "org.scala-sbt" %% "io" % "0.13.8"
+  val sbtIO = "org.scala-sbt" %% "io" % "1.0.0-M9"
 
   val scalaAsync = "org.scala-lang.modules" %% "scala-async" % "0.9.2"
   val scalaJava8 = "org.scala-lang.modules" %% "scala-java8-compat" % "0.7.0"
